+++ tests/testthat/test-clinicaltrials_gov_version_max_age.R
@@ -1,0 +1,14 @@
+test_that("ClinicalTrials.gov maximum age downloads correctly", {
+    version <- clinicaltrials_gov_version("NCT02110043", 1)
+    if (version[1] == "Error") {
+        expect_equal(
+            version[1],
+            "Error"
+        ) 
+    } else {
+        expect_equal(
+            version$max_age,
+            "80"
+        )
+    }
+})